image_search/requirements.txt
ipyplot == 1.1.0 
ftfy ==  4.0.11
python-dotenv == 1.0.1
scikit-learn == 1.4.1.post1 
chromadb == 0.4.24
matplotlib == 3.8.3
streamlit == 1.32.0
watchdog == 4.0.0
pillow == 10.2.0